artifacts/io.github.rally-tavern/python-fastapi-sso-starter/templates/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.27
alembic>=1.13.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
authlib>=1.3.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic[email]>=2.6.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
